feat: moved routes from main to their respective modules
